Restructured code to not use global player_alive variable
When an enemy is spawned in the main scene, it has a set_player() function that is given the player's instance as a parameter. In the enemy script, the player's death signal is connected to the enemy's _on_player_death() function, so that function will react when the palyer's death signal is emmited.